+++ test/requests/sql/test_get_database.rb
@@ -25,12 +25,11 @@ class TestGetDatabase < Minitest::Test
   end
 
   def test_get_database_exception
-    response = -> { fail Exception.new('mocked exception') }
+    response = proc { raise MsRestAzure::AzureOperationError.new(nil, nil, 'error' => { 'message' => 'mocked exception' }) }
     @token_provider.stub :get_authentication_header, response do
       assert_raises Exception do
         @service.get_database('fog-test-rg', 'fog-test-server-name', 'fog-test-database-name')
       end
     end
   end
-
 end