+++ test/requests/sql/test_list_databases.rb
@@ -25,7 +25,7 @@ class TestListDatabases < Minitest::Test
   end
 
   def test_list_databases_exception
-    response = -> { fail Exception.new('mocked exception') }
+    response = proc { raise MsRestAzure::AzureOperationError.new(nil, nil, 'error' => { 'message' => 'mocked exception' }) }
     @token_provider.stub :get_authentication_header, response do
       assert_raises Exception do
         @service.list_databases('fog-test-rg', 'fog-test-zone')