+++ test/requests/sql/test_list_sql_servers.rb
@@ -25,7 +25,7 @@ class TestListSqlServers < Minitest::Test
   end
 
   def test_list_sql_servers_exception
-    response = -> { fail Exception.new('mocked exception') }
+    response = proc { raise MsRestAzure::AzureOperationError.new(nil, nil, 'error' => { 'message' => 'mocked exception' }) }
     @token_provider.stub :get_authentication_header, response do
       assert_raises Exception do
         @service.list_sql_servers('fog-test-rg')